Deliver configuration via a configmap
This makes it read-only and any changes made in the middleware and search-
interface would be lost in the next deployment unless the changes was persisted
to the configmap first.

This should be ok as we don't really expect ever to make any changes to their
configurations, and if we do it's just a matter of updating the configmap.

REYOS2-4