Rebuild all images and bumped build-version to 0.3.0
Mainly done to get rid of an inconsistency in the admin-php image (it where
missing a patch package even though php-base was supposed to provide it), but
also to get a fresh rebuild with package updates.

This also means that all the images are back on the same build-version which
makes our versions just that bit less confusing.

No source-updates, so all our application-revisions are the same.